+++ sdk/communication/Azure.Communication.Identity/src/CommunicationIdentityClient.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Threading;
 using System.Threading.Tasks;
 using Azure.Communication.Pipeline;
@@ -203,17 +204,68 @@ namespace Azure.Communication.Identity
         }
 
         /// <summary>Gets a token for a <see cref="CommunicationUserIdentifier"/>.</summary>
-        /// <param name="communicationUser">The <see cref="CommunicationUserIdentifier"/> for whom to get a token.</param>
-        /// <param name="scopes">The scopes that the token should have.</param>
+        /// <param name="getTokenOptions">Options used to get a token for a <see cref="CommunicationUserIdentifier"/>.</param>
         /// <param name="cancellationToken">The cancellation token to use.</param>
         /// <exception cref="RequestFailedException">The server returned an error.</exception>
-        public virtual Response<AccessToken> GetToken(CommunicationUserIdentifier communicationUser, IEnumerable<CommunicationTokenScope> scopes, CancellationToken cancellationToken = default)
+        public virtual Response<AccessToken> GetToken(GetTokenOptions getTokenOptions, CancellationToken cancellationToken = default)
         {
             using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(CommunicationIdentityClient)}.{nameof(GetToken)}");
             scope.Start();
             try
             {
-                Response<CommunicationIdentityAccessToken> response = RestClient.IssueAccessToken(communicationUser.Id, scopes, cancellationToken);
+                int? tokenExpiresInMinutes = getTokenOptions.ExpiresInMinutes.HasValue ?
+                    Convert.ToInt32(getTokenOptions.ExpiresInMinutes.Value.TotalMinutes) :
+                    null;
+
+                Response<CommunicationIdentityAccessToken> response = RestClient
+                    .IssueAccessToken(
+                    getTokenOptions.CommunicationUser.Id,
+                    getTokenOptions.Scopes,
+                    tokenExpiresInMinutes,
+                    cancellationToken);
+
+                return Response.FromValue(new AccessToken(response.Value.Token, response.Value.ExpiresOn), response.GetRawResponse());
+            }
+            catch (Exception ex)
+            {
+                scope.Failed(ex);
+                throw;
+            }
+        }
+
+        /// <summary>Gets a token for a <see cref="CommunicationUserIdentifier"/>.</summary>
+        /// <param name="communicationUser">The <see cref="CommunicationUserIdentifier"/> for whom to get a token.</param>
+        /// <param name="scopes">The scopes that the token should have.</param>
+        /// <param name="cancellationToken">The cancellation token to use.</param>
+        /// <exception cref="RequestFailedException">The server returned an error.</exception>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public virtual Response<AccessToken> GetToken(CommunicationUserIdentifier communicationUser, IEnumerable<CommunicationTokenScope> scopes, CancellationToken cancellationToken = default)
+        {
+            var getTokenOptions = new GetTokenOptions(communicationUser, scopes);
+            return GetToken(getTokenOptions, cancellationToken);
+        }
+
+        /// <summary>Asynchronously gets a token for a <see cref="CommunicationUserIdentifier"/>.</summary>
+        /// <param name="getTokenOptions">Options used to get a token for a <see cref="CommunicationUserIdentifier"/>.</param>
+        /// <param name="cancellationToken">The cancellation token to use.</param>
+        public virtual async Task<Response<AccessToken>> GetTokenAsync(GetTokenOptions getTokenOptions, CancellationToken cancellationToken = default)
+        {
+            using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(CommunicationIdentityClient)}.{nameof(GetToken)}");
+            scope.Start();
+            try
+            {
+                int? tokenExpiresInMinutes = getTokenOptions.ExpiresInMinutes.HasValue ?
+                    Convert.ToInt32(getTokenOptions.ExpiresInMinutes.Value.TotalMinutes) :
+                    null;
+
+                Response<CommunicationIdentityAccessToken> response = await RestClient
+                    .IssueAccessTokenAsync(
+                    getTokenOptions.CommunicationUser.Id,
+                    getTokenOptions.Scopes,
+                    tokenExpiresInMinutes,
+                    cancellationToken)
+                    .ConfigureAwait(false);
+
                 return Response.FromValue(new AccessToken(response.Value.Token, response.Value.ExpiresOn), response.GetRawResponse());
             }
             catch (Exception ex)
@@ -227,20 +279,11 @@ namespace Azure.Communication.Identity
         /// <param name="communicationUser">The <see cref="CommunicationUserIdentifier"/> for whom to get a token.</param>
         /// <param name="scopes">The scopes that the token should have.</param>
         /// <param name="cancellationToken">The cancellation token to use.</param>
+        [EditorBrowsable(EditorBrowsableState.Never)]
         public virtual async Task<Response<AccessToken>> GetTokenAsync(CommunicationUserIdentifier communicationUser, IEnumerable<CommunicationTokenScope> scopes, CancellationToken cancellationToken = default)
         {
-            using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(CommunicationIdentityClient)}.{nameof(GetToken)}");
-            scope.Start();
-            try
-            {
-                Response<CommunicationIdentityAccessToken> response = await RestClient.IssueAccessTokenAsync(communicationUser.Id, scopes, cancellationToken).ConfigureAwait(false);
-                return Response.FromValue(new AccessToken(response.Value.Token, response.Value.ExpiresOn), response.GetRawResponse());
-            }
-            catch (Exception ex)
-            {
-                scope.Failed(ex);
-                throw;
-            }
+            var getTokenOptions = new GetTokenOptions(communicationUser, scopes);
+            return await GetTokenAsync(getTokenOptions, cancellationToken).ConfigureAwait(false);
         }
 
         /// <summary>Revokes all the tokens created for a user.</summary>